+++ xfa/fxfa/cxfa_textlayout.h
@@ -111,8 +111,7 @@ class CXFA_TextLayout final : public cppgc::GarbageCollected<CXFA_TextLayout> {
     float fLastPos = 0;
     float fStartLineOffset = 0;
     size_t nCharIdx = 0;
-    // TODO(thestig): Make this size_t?
-    int32_t iTotalLines = -1;
+    size_t iTotalLines = -1;
     UnownedPtr<const CFX_XMLNode> pXMLNode;
     RetainPtr<CFX_CSSComputedStyle> pParentStyle;
     cppgc::Member<CXFA_Node> pNode;
